restructuring and standardising tests
